+++ src/lib/components/ui/dropdown-menu/index.ts
@@ -1,0 +1,16 @@
+import Root from './dropdown-menu.svelte';
+import Trigger from './dropdown-menu-trigger.svelte';
+import Content from './dropdown-menu-content.svelte';
+import Item from './dropdown-menu-item.svelte';
+
+export {
+    Root,
+    Trigger,
+    Content,
+    Item,
+    //
+    Root as DropdownMenu,
+    Trigger as DropdownMenuTrigger,
+    Content as DropdownMenuContent,
+    Item as DropdownMenuItem
+};