+++ src/lib/index.ts
@@ -1,9 +1,9 @@
 // place files you want to import through the `$lib` alias in this folder.
 
 export let seo = {
-	title: 'Svelte Startup Template',
+	title: 'your finance assistant',
 	description:
-		'Svelte Startup Template is a template for building Startup Applications. It is built with SvelteKit, TailwindCSS, and Svelte Animations',
+		'your finance assistant is a template for building Startup Applications. It is built with SvelteKit, TailwindCSS, and Svelte Animations',
 	image: 'https://i.pinimg.com/736x/85/9a/92/859a92a2629f912010a0a72270aefedc.jpg',
 	twitter: 'SEO twitter',
 	url: 'https://startup-sve.vercel.app',